+++ inc/Helpers/ScreenImplementation.php
@@ -37,8 +37,8 @@ class ScreenImplementation {
 	 * @return boolean
 	 */
 	public function in_block_editor_admin(): bool {
-		if ( function_exists( 'get_current_screen' ) ) {
-			$screen = get_current_screen();
+		if ( function_exists( '\get_current_screen' ) ) {
+			$screen = \get_current_screen();
 
 			if ( $screen instanceof \WP_Screen ) {
 				return $screen->is_block_editor();
@@ -92,7 +92,7 @@ class ScreenImplementation {
 			$post_type = $typenow;
 		}
 
-		if ( empty( $post_type ) && function_exists( 'get_current_screen' ) ) {
+		if ( empty( $post_type ) && function_exists( '\get_current_screen' ) ) {
 			$screen = \get_current_screen();
 
 			if ( $screen ) {
@@ -101,9 +101,9 @@ class ScreenImplementation {
 		}
 
 		// phpcs:ignore: WordPress.Security.NonceVerification.Recommended
-		if ( empty( $post_type ) && isset( $_REQUEST['post'] ) && ! empty( $_REQUEST['post'] ) && function_exists( 'get_post_type' ) ) {
+		if ( empty( $post_type ) && isset( $_REQUEST['post'] ) && ! empty( $_REQUEST['post'] ) && function_exists( '\get_post_type' ) ) {
 			// phpcs:ignore: WordPress.Security.NonceVerification.Recommended
-			$get_post_type = get_post_type( (int) $_REQUEST['post'] );
+			$get_post_type = \get_post_type( (int) $_REQUEST['post'] );
 			$post_type = $get_post_type;
 		}